Update useMutation usage in addNewPost
- transfer addNewPost.tsx to components folder
- use project react query structure
